Fixed null pointer error on the jeevesAppContext while destroying object. It would occur when GN failed to startup correctly.

+++ core/src/main/java/jeeves/config/springutil/JeevesContextLoaderListener.java
@@ -105,7 +105,9 @@ public class JeevesContextLoaderListener implements ServletContextListener {
          */
         JeevesApplicationContext jeevesAppContext =
             (JeevesApplicationContext) servletContext.getAttribute(User.NODE_APPLICATION_CONTEXT_KEY);
-        jeevesAppContext.destroy();
+        if (jeevesAppContext != null) {
+            jeevesAppContext.destroy();
+        }
 
         Log.info(Log.ENGINE, "Destroying the parent appContext");
         parentAppContext.destroy();
